+++ core/src/proxy/health-check/index.ts
@@ -1,1 +1,7 @@
+export type { HealthCheckManagerParams } from "./health-check-manager";
+export { HealthCheckManager } from "./health-check-manager";
 export type { HealthCheckStatus } from "./health-check-status";
+export type {
+  HealthCheckStatusChanged,
+  HealthCheckStatusChangedHandler,
+} from "./health-check-status-changed";