+++ core/src/proxy/proxy-factory.ts
@@ -1,7 +1,6 @@
-import { AmazonConnectConfig } from "../amazon-connect-config";
 import { AmazonConnectProvider } from "../provider";
 import { Proxy } from "./proxy";
 
-export type ProxyFactory<TConfig extends AmazonConnectConfig> = (
-  provider: AmazonConnectProvider<TConfig>,
+export type ProxyFactory<TProvider extends AmazonConnectProvider> = (
+  provider: TProvider,
 ) => Proxy;