+++ user/src/routes.ts
@@ -1,3 +1,8 @@
 export enum UserRoutes {
   getLanguage = "getLanguage",
 }
+
+export enum SettingsRoutes {
+  getLanguage = "getLanguage",
+  setLanguage = "setLanguage",
+}